instadeploy: make sure we only send one http result
The previous commit fixed an issue where sending data to the client
from the server needed to be utf-8 encoded. This caused an exception
when sending the HTTP result, which was caught, and another exception
was thrown, causing 200 OK headers to be sent first, then within the
body a HTTP/1.0 500 error. Very confusing to debug.

This commit fixes that, by returning only one tuple from the
`process_request` method, and using that in the `do_POST` method.